pyproject.toml
[project]
name = "geoguessr-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "browser-use>=0.1.40",
    "lmnr[all]>=0.4.62",
    "openai>=1.64.0",
    "pillow>=11.1.0",
    "tabulate>=0.9.0",
    "types-tabulate>=0.9.0.20241207",
]

[tool.uv]
dev-dependencies = [
    "mypy>=1.15.0",
    "ruff>=0.9.7",
]
